+++ src/models/EntryType.php
@@ -71,6 +71,12 @@ class EntryType extends Model implements FieldLayoutProviderInterface
     public ?string $titleFormat = null;
 
     /**
+     * @var bool Whether to show the Slug field
+     * @since 5.0.0
+     */
+    public bool $showSlugField = true;
+
+    /**
      * @var string Slug translation method
      * @phpstan-var Field::TRANSLATION_METHOD_NONE|Field::TRANSLATION_METHOD_SITE|Field::TRANSLATION_METHOD_SITE_GROUP|Field::TRANSLATION_METHOD_LANGUAGE|Field::TRANSLATION_METHOD_CUSTOM
      * @since 4.5.0
@@ -137,6 +143,7 @@ class EntryType extends Model implements FieldLayoutProviderInterface
             'name' => Craft::t('app', 'Name'),
             'titleFormat' => Craft::t('app', 'Title Format'),
             'showStatusField' => Craft::t('app', 'Show the Status field'),
+            'showSlugField' => Craft::t('app', 'Show the Slug field'),
         ];
     }
 
@@ -245,6 +252,7 @@ class EntryType extends Model implements FieldLayoutProviderInterface
             'titleTranslationMethod' => $this->titleTranslationMethod,
             'titleTranslationKeyFormat' => $this->titleTranslationKeyFormat,
             'titleFormat' => $this->titleFormat,
+            'showSlugField' => $this->showSlugField,
             'slugTranslationMethod' => $this->slugTranslationMethod,
             'slugTranslationKeyFormat' => $this->slugTranslationKeyFormat,
             'showStatusField' => $this->showStatusField,