+++ src/records/EntryType.php
@@ -24,6 +24,7 @@ use yii2tech\ar\softdelete\SoftDeleteBehavior;
  * @property string $titleTranslationMethod Title translation method
  * @property string|null $titleTranslationKeyFormat Title translation key format
  * @property string|null $titleFormat Title format
+ * @property bool $showSlugField Whether to show the Slug field
  * @property string $slugTranslationMethod Slug translation method
  * @property string|null $slugTranslationKeyFormat Slug translation key format
  * @property bool $showStatusField Whether to show the Status field