+++ src/services/Entries.php
@@ -1231,12 +1231,16 @@ SQL)->execute();
             ->where(['dateDeleted' => null]);
 
         // todo: remove after the next breakpoint
-        if (Craft::$app->getDb()->columnExists(Table::ENTRYTYPES, 'slugTranslationMethod')) {
+        $db = Craft::$app->getDb();
+        if ($db->columnExists(Table::ENTRYTYPES, 'slugTranslationMethod')) {
             $query->addSelect([
                 'slugTranslationMethod',
                 'slugTranslationKeyFormat',
                 'showStatusField',
             ]);
+        }
+        if ($db->columnExists(Table::ENTRYTYPES, 'showSlugField')) {
+            $query->addSelect('showSlugField');
         }
 
         return $query;
@@ -1377,6 +1381,7 @@ SQL)->execute();
             $entryTypeRecord->titleTranslationMethod = $data['titleTranslationMethod'] ?? '';
             $entryTypeRecord->titleTranslationKeyFormat = $data['titleTranslationKeyFormat'] ?? null;
             $entryTypeRecord->titleFormat = $data['titleFormat'];
+            $entryTypeRecord->showSlugField = $data['showSlugField'] ?? true;
             $entryTypeRecord->slugTranslationMethod = $data['slugTranslationMethod'] ?? Field::TRANSLATION_METHOD_SITE;
             $entryTypeRecord->slugTranslationKeyFormat = $data['slugTranslationKeyFormat'] ?? null;
             $entryTypeRecord->showStatusField = $data['showStatusField'] ?? true;